feat(connections): Virtualize connection list & refactor hooks
- Replace mapped list with react-window FixedSizeList for O(1) rendering
- Move useContainerDimensions and useWindowWidth to src/common/hooks.ts
- Refactor Connections page layout to flex column for proper list containment
- Remove heavy entry/exit animations for list items
