+++ src/common/hooks.ts
@@ -10,3 +10,33 @@ export function useDelay(ms: number) {
 
     return isReady;
 }
+
+export function useContainerDimensions(ref: React.RefObject<HTMLDivElement | null>) {
+    const [dimensions, setDimensions] = useState({ width: 0, height: 0 });
+
+    useEffect(() => {
+        if (!ref.current) return;
+
+        const observer = new ResizeObserver(entries => {
+            if (!entries || entries.length === 0) return;
+            const { width, height } = entries[0].contentRect;
+            setDimensions({ width, height });
+        });
+
+        observer.observe(ref.current);
+        return () => observer.disconnect();
+    }, [ref]);
+
+    return dimensions;
+}
+
+export function useWindowWidth() {
+    const [width, setWidth] = useState(typeof window !== 'undefined' ? window.innerWidth : 0);
+
+    useEffect(() => {
+        const handleResize = () => setWidth(window.innerWidth);
+        window.addEventListener('resize', handleResize);
+        return () => window.removeEventListener('resize', handleResize);
+    }, []);
+    return width;
+}
